+++ src/components/challengeView/scene.ts
@@ -1,5 +1,5 @@
-import { adaptURL } from "../../../scriptLoader";
-import { Challenge } from "../../../staticData/challenges";
+import { adaptURL } from "../../scriptLoader";
+import { Challenge } from "../../staticData/challenges";
 
 class Scene {
     iframe(): HTMLIFrameElement {
@@ -42,11 +42,11 @@ class Scene {
         })
     }
 
-    listenToIframeMessages() {      
+    listenToIframeMessages() {
         window.addEventListener("message", (event) => {
             // exercises post error messages in the form  { tipo: "error", error: object }
             // where object can be any error or { name: "ActividadError", message: "description"}
-            if(event.data.tipo === "error")
+            if (event.data.tipo === "error")
                 console.log(`Pilasweb execution ended with error: ${JSON.stringify(event.data.error)}`)
         })
     }
@@ -106,6 +106,17 @@ class Scene {
     
             comportamiento;
           `)
+    }
+
+    evaluateExpression(expression: string) {
+        return this.eval(`
+        try {
+          var value = pilas.escena_actual().automata.${expression}
+        } catch (e) {
+          pilas.escena_actual().errorHandler.handle(e);
+        }
+
+        value`)
     }
 }
